requirements.txt
openai==0.28
Pillow==10.1.0
Requests==2.31.0
streamlit==1.28.2
transformers==4.34.1
torch==2.0.1
pytest==7.4.0
